+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 2.2.1
-* Fixed Australian 2021 grand prix name.
+* Updated Australia track segment name and track name
+* Updated Italy track segment name and description
 
 ## 2.2.0
 * Added trackSegment.eventSeason core metadata.